requirements.txt
advent-of-code-data==0.9.7
numpy==1.19.4
python-dotenv==0.15.0
